+++ azure-websockets/.env.example
@@ -1,14 +1,20 @@
 DISCORD_BOT_TOKEN=
-DISCORD_BOT_CLIENT_ID=
+DISCORD_CLIENT_ID=
 
 LEETCODE_RELEVANT_TAGS=
 
 MAL_CLIENT_ID=
 MAL_CLIENT_SECRET=
+MAL_OAUTH_REDIRECT_URI=
+DATABASE_MAL_OAUTH_TABLE=
 
-AZURE_SQL_USERNAME=
-AZURE_SQL_PASSWORD=
+COSMOS_ENDPOINT=
+COSMOS_KEY=
+COSMOS_DATABASE=
 
-AZURE_MAL_OAUTH_URL=
+REDIS_HOST=
+REDIS_PRIMARY_ACCESS_KEY=
 
 PUPPETEER_REMOTE_BROWSER_KEY=
+
+PORT=4000
